pr fixes part1
